Distributed components for sections. Separated the styles by components in header.
