+++ src/components/App.jsx
@@ -1,14 +1,11 @@
-import React from 'react'
 import { LangProvider } from '../context/SiteContext'
-import { UI } from './UI'
+import { UI } from '../UI'
 
 const App = () => {
   return (
-    <>
       <LangProvider>
         <UI />
       </LangProvider>
-    </>
   )
 }
