+++ src/components/Header/LogoContainer.jsx
@@ -1,0 +1,9 @@
+const LogoContainer = ({children}) => {
+  return (
+    <aside>
+      {children}
+    </aside>
+  )
+}
+
+export { LogoContainer }
