+++ src/components/Header/MenuContainer.jsx
@@ -1,0 +1,9 @@
+const MenuContainer = ({children}) => {
+  return (
+    <nav>
+      {children}
+    </nav>
+  )
+}
+
+export { MenuContainer }
