+++ src/components/MainContainer.jsx
@@ -1,0 +1,9 @@
+const MainContainer = (props) => {
+  return (
+    <main className="MainContainer">
+      {props.children}
+    </main>
+  )
+}
+
+export { MainContainer }
