+++ src/styles/BurgerButton.component.css
@@ -1,0 +1,18 @@
+.BurgerButton {
+  cursor: pointer;
+  width: 48px;
+  height: 36px;
+  background-color: var(--brand-color);
+  border-radius: 6px;
+  color: var(--white-color);
+  text-align: center;
+}
+
+.BurgerButton.active {
+  background-color: var(--brand-color-active);
+}
+
+.BurgerButton svg {
+  fill: var(--white-color);
+  font-size: 36px;
+}
