+++ src/styles/Logo.component.css
@@ -1,0 +1,18 @@
+.Logo {
+  height: 56px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 3.2rem;
+  font-weight: 700;
+  cursor: pointer;
+}
+
+.Logo img {
+  width: 48px;
+  height: 48px;
+}
+
+.Logo span {
+  display: none;
+}
